Replaced types of length from unsigned int to size_t.

+++ src/msgpack/sbuffer.h
@@ -64,7 +64,7 @@ static inline void msgpack_sbuffer_free(msgpack_sbuffer* sbuf)
 #define MSGPACK_SBUFFER_INIT_SIZE 8192
 #endif
 
-static inline int msgpack_sbuffer_write(void* data, const char* buf, unsigned int len)
+static inline int msgpack_sbuffer_write(void* data, const char* buf, size_t len)
 {
 	msgpack_sbuffer* sbuf = (msgpack_sbuffer*)data;
